Fixed bug with placeholders (double %% after escaping the message)
if getMessage(...) escapes percentage to double char (%%), then placeholders does not work.
We need to unescape them back before call to PlaceholderAPI.setPlaceholders(...)

The concepts is that we will remember that message was escaped % to %%.
And the method prepareMessage(...) will unescape it again at the end if it was escaped.